backend/requirements.txt
python-dotenv>=1.0.0
langgraph
langchain-core
langchain-openai
uvicorn
fastapi
